LayoutTests/fast/scrolling/scroll-anchoring-constraints-expected.txt

PASS window.scrollY is 0
PASS successfullyParsed is true

TEST COMPLETE
